Bugfix: consider root-relative paths when fingerprinting
This remediates a correctness issue where if a file was MOVED BUT NOT CHANGED
its content hash would be invariant and so the code movement wouldn't impact the
fingerprint of target(s) containing the file.

For instance:
 src/a/foo.lib -> src/b/foo.lib

This has clear semantic consequences - it changes the classpath entry of
`foo.lib` - but the content hash of the file has not changed just from the
move. In order to correctly detect this, we have to consider the path to the
file as well as its contents.