Fix plan to work on the buildgraph
The plan and prep APIs originally functioned on the buildgraph, but before
Maxwell got support for recursively including diffs it really wasn't possible to
implement as originally designed.

Now that Maxwell has support for recursive diff inclusion and we need these
codepaths to work as documented for the new manifest and rule prep rewriting
games, just overhaul it to do the obvious thing.

Note that `diff/without-diff` now DROPS a diffing wrapper. We need to be using
`diff/empty-diff` instead, which clears the associated diff of a wrapper.