Enable Katamari to copy "products" out to the target tree
A sticking point for Katamari is the distinction between an intermediary build
product with which the user really has no business such as a bunch of classfiles
and a "final" product which a user will want to directly consume such as
a jarfile.

This patch introduces the concept of a "product" of a target - being an array of
paths naming files which could for instance be copied to Katamari's configured
target directory once a build completes. This allows for some distinction
between intermediary products which could participate in `:paths` and "terminal"
products which a user may choose to copy around or otherwise use.